+++ pkg/report/default.go
@@ -4,6 +4,7 @@ import (
 	"github.com/cyberark/conjur-inspect/pkg/check"
 	"github.com/cyberark/conjur-inspect/pkg/checks"
 	"github.com/cyberark/conjur-inspect/pkg/checks/disk"
+	"github.com/cyberark/conjur-inspect/pkg/container"
 )
 
 // NewDefaultReport returns a report containing the standard inspection checks
@@ -62,10 +63,23 @@ func defaultReportSections() []Section {
 			},
 		},
 		{
-			Title: "Container Runtime",
+			Title: "Container",
 			Checks: []check.Check{
-				&checks.Docker{},
-				&checks.Podman{},
+				// Runtime
+				&checks.ContainerRuntime{
+					Provider: &container.DockerProvider{},
+				},
+				&checks.ContainerRuntime{
+					Provider: &container.PodmanProvider{},
+				},
+
+				// Container inspect
+				&checks.ContainerInspect{
+					Provider: &container.DockerProvider{},
+				},
+				&checks.ContainerInspect{
+					Provider: &container.PodmanProvider{},
+				},
 			},
 		},
 		{